Fix plugin dispatch build wiring
